Revert msm camera_v2 code to Motorola's release.
Something in the numerous changes merged in was causing the front
facing camera to not take still pictures properly. I'm not in the
mood to go through commit by commit to find the cause, and these
changes don't really improve end user experience, so who cares.
Thus, I'm just using the brute force approach of reverting the
entire camera directory for now.